src/requirements.txt
django
ipython
psycopg2-binary
factory_boy
